make immortal mode configurable in runtime
