+++ conf/examples/silver-empty-schema.conf
@@ -9,7 +9,8 @@
         "LogFile": "",
         "LogFileMaxSizeMb": 200,
         "PidFile": "${ServiceName}.pid",
-        "CrashLogFile": "${ServiceName}.crashlog"
+        "CrashLogFile": "${ServiceName}.crashlog",
+        "AttemptRestartOnPanic": false
     },
     "Services": [
         {
